+++ buildSrc/CHANGELOG.md
@@ -9,9 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - A `SupportedVariants` enumeration type.
-- A `jvm.variants` build configuration property.
+- `jvm.variants` and `konan.variants` build configuration properties.
 - A `jvmName` DSL helper method that works with the `SupportedVariants` enumeration type.
-- A `jvmMain` and `jvmTest` DSL helper methods taking a `Named` target.
+- `jvmMain` and `jvmTest` DSL helper methods taking a `Named` target.
 
 ### Fixed